Code Refactoring
