+++ learn-designpatterns-in-csharp/Source/Behavioral/Memento/EmployeeManagerDemo/CareTakers/CommandManager.cs
@@ -1,0 +1,50 @@
+﻿using EmployeeManagerDemo.Commands;
+using EmployeeManagerDemo.Mementos;
+
+namespace EmployeeManagerDemo.CareTakers
+{
+    /// <summary>
+    /// CareTaker (and Invoker from the Command pattern)
+    /// </summary>
+    public class CommandManager
+    {
+        private readonly Stack<AddEmployeeToManagerListMemento> _mementos = new();
+        private AddEmployeeToManagerList? _command;
+
+        public void Invoke(AddEmployeeToManagerList command)
+        {
+            // if the command has not been stored yet, store it - we will
+            // reuse it instead of storing different instances
+
+            if (_command == null)
+            {
+                _command = command;
+            }
+
+            if (command.CanExecute())
+            {
+                command.Execute();
+                _mementos.Push(command.CreateMemento());
+            }
+        }
+
+        public void Undo()
+        {
+            if (_mementos.Any())
+            {
+                _command?.RestoreMemento(_mementos.Pop());
+                _command?.Undo();
+            }
+        }
+
+        public void UndoAll()
+        {
+            while (_mementos.Any())
+            {
+                _command?.RestoreMemento(_mementos.Pop());
+                _command?.Undo();
+            }
+        }
+    }
+
+}
