+++ learn-designpatterns-in-csharp/Source/Behavioral/Memento/EmployeeManagerDemo/Commands/ICommand.cs
@@ -1,7 +1,4 @@
-﻿using EmployeeManagerDemo.Mementos;
-using EmployeeManagerDemo.Models;
-
-namespace EmployeeManagerDemo.Commands
+﻿namespace EmployeeManagerDemo.Commands
 {
 
     /// <summary>
@@ -10,124 +7,10 @@ namespace EmployeeManagerDemo.Commands
     public interface ICommand
     {
         void Execute();
+
         bool CanExecute();
+
         void Undo();
-    }
-
-    /// <summary>
-    /// ConcreteCommand
-    /// </summary>
-    public class AddEmployeeToManagerList : ICommand
-    {
-        private readonly IEmployeeManagerRepository _employeeManagerRepository;
-        private int _managerId;
-        private Employee? _employee;
-
-        public AddEmployeeToManagerList(
-            IEmployeeManagerRepository employeeManagerRepository,
-            int managerId,
-            Employee? employee)
-        {
-            _employeeManagerRepository = employeeManagerRepository;
-            _managerId = managerId;
-            _employee = employee;
-        }
-
-        public AddEmployeeToManagerListMemento CreateMemento()
-        {
-            return new AddEmployeeToManagerListMemento(_managerId, _employee);
-        }
-
-        public void RestoreMemento(AddEmployeeToManagerListMemento memento)
-        {
-            _managerId = memento.ManagerId;
-            _employee = memento.Employee;
-        }
-
-        public bool CanExecute()
-        {
-            // potentially check against business/technical rules that might block execution
-
-            // employee shouldn't be null
-            if (_employee == null)
-            {
-                return false;
-            }
-
-            // employee shouldn't be on the manager's list already
-            if (_employeeManagerRepository.HasEmployee(_managerId, _employee.Id))
-            {
-                return false;
-            }
-
-            // other potential rule(s): ensure that an employee can only be added to  
-            // one manager's list at the same time, etc.
-            return true;
-        }
-
-        public void Execute()
-        {
-            if (_employee == null)
-            {
-                return;
-            }
-
-            _employeeManagerRepository.AddEmployee(_managerId, _employee);
-        }
-
-        public void Undo()
-        {
-            if (_employee == null)
-            {
-                return;
-            }
-
-            _employeeManagerRepository.RemoveEmployee(_managerId, _employee);
-        }
-    }
-
-    /// <summary>
-    /// CareTaker (and Invoker from the Command pattern)
-    /// </summary>
-    public class CommandManager
-    {
-        private readonly Stack<AddEmployeeToManagerListMemento> _mementos = new();
-        private AddEmployeeToManagerList? _command;
-
-        public void Invoke(AddEmployeeToManagerList command)
-        {
-            // if the command has not been stored yet, store it - we will
-            // reuse it instead of storing different instances
-
-            if (_command == null)
-            {
-                _command = command;
-            }
-
-            if (command.CanExecute())
-            {
-                command.Execute();
-                _mementos.Push(command.CreateMemento());
-            }
-        }
-
-        public void Undo()
-        {
-            if (_mementos.Any())
-            {
-                _command?.RestoreMemento(_mementos.Pop());
-                _command?.Undo();
-            }
-        }
-
-        public void UndoAll()
-        {
-            while (_mementos.Any())
-            {
-                _command?.RestoreMemento(_mementos.Pop());
-                _command?.Undo();
-            }
-        }
     }
 
 }
